64bit bug fix

+++ envvar-cmdline.cpp
@@ -119,7 +119,7 @@ JNIEXPORT jstring JNICALL Java_org_jvnet_winp_Native_getCmdLineAndEnvVars(
 JNIEXPORT jint JNICALL Java_org_jvnet_winp_Native_getProcessId(JNIEnv* pEnv, jclass clazz, jint handle) {
 	HANDLE hProcess = (HANDLE)handle;
 	PROCESS_BASIC_INFORMATION ProcInfo;
-	ULONG _;
+	SIZE_T _;
 	if(!NT_SUCCESS(ZwQueryInformationProcess(hProcess, ProcessBasicInformation, &ProcInfo, sizeof(ProcInfo), &_))) {
 		reportError(pEnv,"Failed to ZWQueryInformationProcess");
 		return -1;
